Create new model files
