--- Pear/Models/UserSocialMedia.swift
@@ -1,16 +1,0 @@
-//
-//  UserSocialMedia.swift
-//  Pear
-//
-//  Created by Lucy Xu on 3/6/21.
-//  Copyright © 2021 cuappdev. All rights reserved.
-//
-
-import Foundation
-
-struct UserSocialMedia: Codable {
-
-    let facebook: String?
-    let instagram: String?
-
-}
