+++ Pear/Models/v1/LocationV1.swift
@@ -1,0 +1,16 @@
+//
+//  LocationV1.swift
+//  Pear
+//
+//  Created by Lucy Xu on 9/1/21.
+//  Copyright © 2021 cuappdev. All rights reserved.
+//
+
+import Foundation
+
+struct LocationV1: Codable {
+
+    let area: String
+    let name: String
+
+}
